--- packages/slider/src/Sliderhandle.tsx
@@ -1,19 +1,0 @@
-import React from "react"
-import "./styles/index.css"
-import Slider, { SliderTooltip } from "rc-slider"
-const { Handle } = Slider
-
-export const handle = (props: any) => {
-  const { value, dragging, index, ...restProps } = props
-  return (
-    <SliderTooltip
-      prefixCls="rc-slider-tooltip"
-      overlay={`${value}`}
-      visible={dragging}
-      placement="top"
-      key={index}
-    >
-      <Handle value={value} {...restProps} />
-    </SliderTooltip>
-  )
-}
