+++ packages/slider/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./Slider"
-export * from "./Sliderhandle"
